# http/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Centrifuge OS Node API",
        "title": "Centrifuge OS Node API",
        "contact": {
            "name": "Centrifuge",
            "url": "https://github.com/centrifuge/pod",
            "email": "hello@centrifuge.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "3.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/v2/documents/{document_id}/roles": {
            "post": {
                "description": "Adds a new role to the document.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Adds a new role to the document.",
                "operationId": "add_role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add Role Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2.AddRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2.Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/documents/{document_id}/roles/{role_id}": {
            "get": {
                "description": "Returns the role associated with the role ID in the document.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Returns the role associated with the role ID in the document.",
                "operationId": "get_role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2.Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates an existing role on the document.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Updates an existing role on the document.",
                "operationId": "update_role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document Identifier",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Role Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2.UpdateRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2.Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/v2/jobs/{job_id}": {
            "get": {
                "description": "Returns the details of a given Job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Returns the details of a given Job.",
                "operationId": "get_job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex encoded centrifuge ID of the account for the intended API action",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hex encoded Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gocelery.Task": {
            "type": "object",
            "properties": {
                "args": {
                    "description": "arguments passed to this task",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "delay": {
                    "description": "delay until ready to be run",
                    "type": "string"
                },
                "error": {
                    "description": "error after task run",
                    "type": "string"
                },
                "result": {
                    "description": "result after the task run",
                    "type": "object"
                },
                "runnerFuncs": {
                    "description": "name of the runnerFuncs",
                    "type": "string"
                },
                "tries": {
                    "description": "number of times task was run.",
                    "type": "integer"
                }
            }
        },
...
        "httputils.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "v2.AddRole": {
            "type": "object",
            "properties": {
                "collaborators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "key": {
                    "description": "Key is either hex encoded 32 byte ID or string label.\nString label is used as a preimage to sha256 for 32 byte hash.",
                    "type": "string"
                }
            }
        },
...
        "v2.Job": {
            "type": "object",
            "properties": {
                "JobID": {
                    "description": "Job Identifier",
                    "type": "string"
                },
                "desc": {
                    "description": "description of the Job",
                    "type": "string"
                },
                "finished": {
                    "description": "job status",
                    "type": "boolean"
                },
                "finished_at": {
                    "description": "Job finished at. If empty, job is not complete yet",
                    "type": "string"
                },
                "overrides": {
                    "description": "overrides for the Job",
                    "type": "object",
                    "additionalProperties": true
                },
                "runner": {
                    "description": "name of the Runner",
                    "type": "string"
                },
                "tasks": {
                    "description": "list of tasks ran under this Job",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gocelery.Task"
                    }
                },
                "valid_until": {
                    "description": "validity of the job",
                    "type": "string"
                }
            }
        },
...
        "v2.Role": {
            "type": "object",
            "properties": {
                "collaborators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "v2.UpdateRole": {
            "type": "object",
            "properties": {
                "collaborators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}